support for 64-bit

+++ services/transient_task/src/delay_suspend_info_ex.cpp
@@ -30,7 +30,7 @@ DelaySuspendInfoEx::DelaySuspendInfoEx(const int32_t& pid, const int32_t& reques
 int32_t DelaySuspendInfoEx::GetRemainDelayTime()
 {
     int64_t spendTime = (baseTime_ > 0) ? (spendTime_ + (TimeProvider::GetCurrentTime() - baseTime_)) : spendTime_;
-    int32_t remainTime = GetActualDelayTime() - spendTime;
+    int32_t remainTime = GetActualDelayTime() - (int32_t)spendTime;
     return (remainTime < 0) ? 0 : remainTime;
 }
